// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
type ServicepageDocumentDataSlicesSlice =
  | AboutSlice
  | MethodeTravailleSlice
  | OurClientsSlice
  | ChiffresSlice
  | OurSectorsSlice
  | ServicesCardsSlice
  | FaqSlice
  | FonctionnalitesSlice
  | SpacerSlice
  | ServiceCallToActionFormSlice
  | ServiceReviewsSlice
  | ServicesPortfolioSlice
  | CallToActionSlice
  | ServiceAboutSlice
  | AdvantagesCardsSlice
  | ServiceHeroSlice;

/**
 * Content for ServicePage documents
 */
interface ServicepageDocumentData {
  /**
   * Type Service field in *ServicePage*
   *
   * - **Field Type**: Content Relationship
   * - **Placeholder**: *None*
   * - **API ID Path**: servicepage.type_service
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#link-content-relationship
   */
  type_service: prismic.ContentRelationshipField<"servicetype">;

  /**
   * Nom Service field in *ServicePage*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: servicepage.nom_service
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  nom_service: prismic.KeyTextField;

  /**
   * Slice Zone field in *ServicePage*
   *
   * - **Field Type**: Slice Zone
   * - **Placeholder**: *None*
   * - **API ID Path**: servicepage.slices[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#slices
   */
  slices: prismic.SliceZone<ServicepageDocumentDataSlicesSlice> /**
   * Meta Title field in *ServicePage*
   *
   * - **Field Type**: Text
   * - **Placeholder**: A title of the page used for social media and search engines
   * - **API ID Path**: servicepage.meta_title
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */;
  meta_title: prismic.KeyTextField;

  /**
   * Meta Description field in *ServicePage*
   *
   * - **Field Type**: Text
   * - **Placeholder**: A brief summary of the page
   * - **API ID Path**: servicepage.meta_description
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  meta_description: prismic.KeyTextField;

  /**
   * Meta Image field in *ServicePage*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: servicepage.meta_image
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/field#image
   */
  meta_image: prismic.ImageField<never>;

  /**
   * meta keywords field in *ServicePage*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: servicepage.meta_keywords
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  meta_keywords: prismic.KeyTextField;
}

/**
 * ServicePage document from Prismic
 *
 * - **API ID**: `servicepage`
 * - **Repeatable**: `true`
 * - **Documentation**: https://prismic.io/docs/custom-types
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type ServicepageDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithUID<
    Simplify<ServicepageDocumentData>,
    "servicepage",
    Lang
  >;
// ...
export interface DropDownMenuSliceDefaultPrimaryServicesTitleItem {
  /**
   * Service Title Link field in *NavMenus → Services DropDown → Primary → Services Title*
   *
   * - **Field Type**: Link
   * - **Placeholder**: *None*
   * - **API ID Path**: drop_down_menu.default.primary.services_title[].service_title_link
   * - **Documentation**: https://prismic.io/docs/field#link-content-relationship
   */
  service_title_link: prismic.LinkField;
}